Refactoring uniforms to use UBOs.
Changes to RrShader, RrMaterial, and additions to the GPU wrapper classes.
Of note is the constant buffer interfaces. It wraps OpenGL UBO's (for now) but a good amount of the Rr_API is being redesigned to take future gfx APIs into account (ie gnm, vulkan, dx11)
All shaders have been refactored to use the newly created structures in RrShader.
RrShader will likely need to go through some refactoring in order for the interface to be well laid out and resistant to user error (which we have a HUGE amount of so it is neeeedeed~~~)